requirements.txt
beautifulsoup4
requests
python-dotenv
lxml
schedule
matplotlib
japanize-matplotlib